Added gRPC service for calculator
A new gRPC service has been added to the project. This service, named CalculatorService, provides a Sum method that can be used by clients to perform addition operations. The implementation includes both client and server interfaces for the CalculatorService along with necessary methods and types. Also, forward compatibility is ensured by embedding UnimplementedCalculatorServiceServer in all implementations of the CalculatorServiceServer interface.