Added protobuf definitions for sum service
Introduced new protobuf definitions for a sum service. This includes `SumRequest` and `SumResponse` message types, each with their respective fields and methods. The `SumRequest` type has two integer fields to hold the numbers to be summed, while the `SumResponse` type contains a single integer field for the result of the operation.