Added functionality for multiple greetings
A new function, doGreetManyTimes, has been added to the client side of the greeting service. This function allows a user to send a single greet request and receive multiple responses. The function logs each response received from the server. Error handling is also included in case of failure during communication or if an error occurs while receiving a response.