Added prime number functionality to calculator server
A new function has been added to the calculator server that calculates prime numbers. This function receives a request, logs it, and then processes the request by dividing the input number by increasing divisors. If the number is divisible, it sends back the divisor as a response and continues with the quotient until no further division is possible.